HEL-409/Afficher_le_donuts_de_l_allocation_de_ressource_v2
