+++ src/pages/etablissement-territorial-sanitaire/[numeroFiness].tsx
@@ -56,172 +56,12 @@ export async function getServerSideProps(context: GetServerSidePropsContext): Pr
       )) as ÉtablissementTerritorialSanitaire;
       const rechercheResult = await rechercheParmiLesEntitésEtÉtablissementsEndpoint(dependencies, numeroFiness, 1);
 
-      const allocationRessources = 
-      [
-        {
-        dateMiseÀJourSource: '2023-01-20',
-        année: 2022,
-        data:
-        [
-          {
-            enveloppe: 'FIR',
-            sousEnveloppe: 'Sanitaire',
-            modeDeDélégation: 'Intervention A',
-            montantNotifié: 1900,
-          },
-          {
-            enveloppe: 'FIR',
-            sousEnveloppe: 'Sanitaire',
-            modeDeDélégation: 'Intervention B',
-            montantNotifié: 50,
-          },
-          {
-            enveloppe: 'FIR',
-            sousEnveloppe: 'Sanitaire',
-            modeDeDélégation: 'Intervention B',
-            montantNotifié: 100,
-          },
-          {
-            enveloppe: 'Forfaits',
-            sousEnveloppe: 'FAU',
-            modeDeDélégation: 'Sans objet',
-            montantNotifié: 570,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'AC',
-            modeDeDélégation: '123',
-            montantNotifié: 6660,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'BASE',
-            montantNotifié: 100,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'JPE_MERRI',
-            montantNotifié: 140,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'TestA',
-            montantNotifié: 190,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'TestB',
-            montantNotifié: 290,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'TestC',
-            montantNotifié: 390,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'TestC',
-            montantNotifié: 490,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'TestD',
-            montantNotifié: 590,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'TestE',
-            montantNotifié: 190,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'TestJ',
-            montantNotifié: 190,
-          },
-          {
-            enveloppe: 'Dotations de soins USLD',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'JPE_MERRI',
-            montantNotifié: 1559,
-          },
-          {
-            enveloppe: 'DAF',
-            sousEnveloppe: 'DAF',
-            modeDeDélégation: 'JPE_MERRI',
-            montantNotifié: 1201,
-          }
-        ]
-      },
-      {
-        dateMiseÀJourSource: '2023-01-20',
-        année: 2019,
-        data:
-        [
-          {
-            enveloppe: 'FIR',
-            sousEnveloppe: 'Sanitaire',
-            modeDeDélégation: 'Intervention A',
-            montantNotifié: 10723000,
-          },
-          {
-            enveloppe: 'FIR',
-            sousEnveloppe: 'Sanitaire',
-            modeDeDélégation: 'Intervention B',
-            montantNotifié: 2000,
-          },
-          {
-            enveloppe: 'FIR',
-            sousEnveloppe: 'Sanitaire',
-            modeDeDélégation: 'Intervention B',
-            montantNotifié: 3000,
-          },
-          {
-            enveloppe: 'Forfaits',
-            sousEnveloppe: 'FAU',
-            modeDeDélégation: 'Sans objet',
-            montantNotifié: 4798 ,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'AC',
-            modeDeDélégation: '123',
-            montantNotifié: 20000,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'BASE',
-            montantNotifié: 17693,
-          },
-          {
-            enveloppe: 'MIGAC',
-            sousEnveloppe: 'MIG',
-            modeDeDélégation: 'JPE_MERRI',
-            montantNotifié:  806 ,
-          }
-        ]
-      }
-    ];
-    
-      return { props: {  
-        
-        // this line is only for fornt test
-        établissementTerritorial: {...établissementTerritorial, 
-          allocationRessources: allocationRessources},
-
-        // use this line after prepare real data
-        //établissementTerritorial,
-        
-        rechercheResult: rechercheResult } };
+      return {
+        props: {
+          établissementTerritorial,
+          rechercheResult: rechercheResult,
+        },
+      };
     } else {
       return { notFound: true };
     }
